+++ server/tsuserver.py
@@ -68,8 +68,8 @@ class TsuserverDR:
         self.release = 4
         self.major_version = 3
         self.minor_version = 1
-        self.segment_version = 'a23'
-        self.internal_version = 'm220305a'
+        self.segment_version = 'a24'
+        self.internal_version = 'm220307a'
         version_string = self.get_version_string()
         self.software = 'TsuserverDR {}'.format(version_string)
         self.version = 'TsuserverDR {} ({})'.format(version_string, self.internal_version)